Resolved range.clear_xxx bugs. Reverted splitting files.
