+++ src/__init__.py
@@ -18,7 +18,5 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-from .main import Workbook, excel2df
-from .range import Range
-from .sheet import Sheet
+from .main import Workbook, excel2df, Sheet, Range
 from .tools import number_to_date, date_to_number
